requirements.txt
Flask==2.0.1
psycopg2-binary==2.9.1
Flask-SQLAlchemy==2.5.1
Flask-Login==0.5.0
Flask-WTF==0.15.1
GeoAlchemy2==0.9.4
# Shapely==2.0.1  # Install separately with: pip install shapely --only-binary :all:
python-dotenv==0.19.0
Flask-Migrate==3.1.0
WTForms==2.3.3
email-validator==1.1.3
Flask-Bcrypt==0.7.1
Werkzeug==2.0.3
SQLAlchemy==1.4.46 
